Updates to Agents to improve performance, accuracy, and poster cropping
Updates to all three scene agents (WayBig, Fagalicious, QueerClick) to ensure that if for any reason the online cropper is unavailable, the backup VBS (windows only) script will fire to crop the poster to display optimally in plex.

GEVI update to better handle titles beginning with "The" or ending with a number "3".  GEVI is tricky in those queries.

AEBN update improves poster accuracy to two nearly identical filenames/related movies.