Enhancements to GEVI
Contains the following enhancements:

I. COLLECTIONS: I did some more work on GEVI and I have tested it on over 100 titles and 20 collections....

As mentioned GEVI has an irritating way of naming titles in series:

        (Studio) - Film - Series (No.) - Year  OR
        (Studio) - Film, Series (No.) - Year  OR
        (Studio) - Series (No.) - Film - Year  OR
        (Studio) - Series (No.), Film - Year  AND
        (Studio) - Film - Year

And in many cases - these will be mixed up as in the case of Lucas Entertainment - Gentlemen collection... a total of 29 films

This meant that to match owt, you had to name them in the exact order they were in GEVI - and in file explorer you would thus not be able to easily identify which titles were missing in a series.

Also, GEVI irritatingly did not have a collection field - unlike AEBN.... So you never had collections matched up when using this agent. And if you had scrapped the title previously in AEBN, you lost this information... You would see this when you tried to browse collections... you would find collections with a total of ZERO movies...

So the new version creates the collection for those which have an ordinal number in them - irrespective of which of the first 4 formats I showed above is used. This means you do not have to rename any of your titles in Explorer.

For format 5 (Studio) - Film (Year) - (Lucas Entertainment) - Executive Authority (2020) on GEVI - the new agent will not create a collection as there is no collection information in the title... To cater for this you can put the collection info like this  (Lucas Entertainment) - Gentlemen 29 - Executive Authority (2020).

For all formats the Agent searches using Short Tile Executive Authority.

When matching it does this to the Short Title - Executive Authority and the Full Title Gentlemen 29 - Executive Authority

Caveat: Collection garnering will only work in Collections with Series Numbers

During the update - if it determines that the Full Title had a collection, then the collection field will be set.

II. ACTORS: As mentioned before I also did some work on this and sorted one small issue last night...
Plex stores three properties per actor:

    Name
    Role: i.e top, bottom, masturbation, etc
    Pictures

IAFD - is searched in 2 ways by Film Title or by Actor as it was before:

These properties are sometimes found fully on IAFD, but in many cases the role or picture can be missing and in some cases GEVI could have more actors or less actors than what is on IAFD... The following is how I display the actors

By Film Title: Role information can only be retrieved when searching by Title and if it is recorded.
If not recorded in IAFD a [yellow] square is displayed under the actors name. this is the default
This process returns 2 lists - unmatched actors and matched actors

By Actor Name: After the film title processing is we attempt to match the unmatched actors:
These will usually be the actors that are shown on GEVI but not on IAFD or those we can't match against an AKA Name
They also default to [yellow] as actors can have different roles in different movies
This process is to try and get their photos if any...
This process also returns 2 lists - unmatched actors and matched actors

Lastly:
At this point any actor left is not recorded on IAFD or is using an AKA name that can not be found in the actor search routine.
The Role will default to a [RED] circle
There will be no photo