Waybig Fix to #84 ; propagate genfunctions.py across agents
Notes from the developer:

When comparing titles, words like Part, Volume, .org, .com etc are stripped from both the website title and the file title - this is done in the normalisation function contained in genfunctions.py

When I was processing the site title, I stripped all the spaces before I ran the normalisation and subsequent comparison.

As the title had Part - it did not strip the word out as there were no spaces in the title saw this was never stripped and the comparison against the file title failed as it was stripped.

For the Studio it was passing an array [-1] this funnily meant that all Studios matched for old titles that had an AT or ON in the title.

I have also removed the line in the synopsis which is off the format Watch Title at studio - just as I do in GEVI

Whilst looking at the gen functions I realised I was replacing ` with ' quotes - however this would leave enquoted words mis balanced if the closing quote was curly. Therefore all quotes now are converted to single straight quotes.