Replaces PGMA's NFO Importer
This release updates and replaces PGMA's NFO Importer.  The Importer will match the actor's from your NFO against IAFD and if it finds a match, returns the actor's photo.

Also, fixes error in GayEmpire - the xpath was not picking up all the search results for some films in the code, but worked when inspecting the page.

Also updates the utils.py, so that the log file displays the Plex Server Version.